pull

+++ Assets/01_Scripts/Scriptable/EntityData.cs
@@ -6,10 +6,10 @@ using UnityEngine;
 public class EntityData : ScriptableObject
 {
     [SerializeField] private string _name;
-    [SerializeField] private int _lifeMax;
-    [SerializeField] private int _damage;
+    [SerializeField] private float _lifeMax;
+    [SerializeField] private float _damage;
 
     public string Name { get => _name; set => _name = value; }
-    public int LifeMax { get => _lifeMax; set => _lifeMax = value; }
-    public int Damage { get => _damage; set => _damage = value; }
+    public float LifeMax { get => _lifeMax; set => _lifeMax = value; }
+    public float Damage { get => _damage; set => _damage = value; }
 }
